+++ src/components/index.js
@@ -1,0 +1,3 @@
+import Nav from "./Nav";
+
+export {Nav};
