+++ src/pages/dashboard/index.js
@@ -1,0 +1,7 @@
+import Albums from './Albums';
+import Comments from './Comments';
+import Photos from './Photos';
+import Posts from './Posts';
+import UserList from './UserList';
+
+export {Albums, Comments, Photos, Posts, UserList};
